api/requirements.txt
tensorflow>=2.0.0
fastapi
uvicorn
python-multipart
pillow
tensorflow-serving-api==2.5.0
matplotlib
numpy
gunicorn==19.7.1